Move ML lib data generator files to util/
